% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{genBaseRPlot}
\alias{genBaseRPlot}
\title{Generate Base R Bar Plot}
\usage{
genBaseRPlot(data)
}
\arguments{
\item{data}{Dataframe containing the average number of lines and file types.}
}
\value{
NULL. The function outputs the plot directly.
}
\description{
Generate Base R Bar Plot
}
