% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{genEcharts}
\alias{genEcharts}
\title{Generate echarts4r Bar Plot}
\usage{
genEcharts(data)
}
\arguments{
\item{data}{Dataframe containing the average number of lines and file types.}
}
\value{
echarts4r plot object.
}
\description{
Generate echarts4r Bar Plot
}
