% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{genGgplot}
\alias{genGgplot}
\title{Generate ggplot Bar Plot}
\usage{
genGgplot(data)
}
\arguments{
\item{data}{Dataframe containing the average number of lines and file types.}
}
\value{
ggplot object.
}
\description{
Generate ggplot Bar Plot
}
