% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{genLatticePlot}
\alias{genLatticePlot}
\title{Generate Lattice Bar Plot}
\usage{
genLatticePlot(data)
}
\arguments{
\item{data}{Dataframe containing the average number of lines and file types.}
}
\value{
lattice plot object.
}
\description{
Generate Lattice Bar Plot
}
