% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotAvgLinesPerFile}
\alias{plotAvgLinesPerFile}
\title{Bar Plot of Average Lines Per File by File Type}
\usage{
plotAvgLinesPerFile(
  dirPath,
  fileTypes = NULL,
  plotLibrary = c("ggplot", "echarts4r", "base", "lattice")
)
}
\arguments{
\item{dirPath}{Character string specifying the path to the directory.}

\item{fileTypes}{Optional character vector specifying which file types to consider.
If NULL (the default), all file types are considered.}

\item{plotLibrary}{Character string specifying which plotting library to use.
Options are "ggplot", "echarts4r", "base", or "lattice". Default is c("ggplot", "echarts4r", "base", "lattice").}
}
\value{
A plot object for ggplot, echarts4r, and lattice. Base R plots are output directly.
}
\description{
Bar Plot of Average Lines Per File by File Type
}
\examples{
\dontrun{
plotAvgLinesPerFile("/path/to/folder", c("txt", "csv", "R"), "lattice")
}

}
